refactor: P2 server split 8/N — extract wiki route domain
Move all /api/wiki/* routes (pages, page GET/PUT/DELETE, graph, stats) and
the /api/l2/raw/upload route out of server.ts into server/routes/wiki.ts as
handleWikiRoutes. Domain-private helpers move with it: WIKI_PAGE_DIRS,
listWikiPagePaths/manifestSourceIdByWikiPath (now take l2DataDir as an
explicit parameter), sanitizeUploadName/uploadExtension. The L2 imports
(getL2Memory, buildWikiGraph, manifest-store, parseFrontmatter) leave
server.ts entirely — nothing outside the wiki domain used them.

The L2 upload route previously sat ~170 lines below the wiki routes; it now
runs from the single delegation point (exact-path match, reordering inert).
server.ts: 2612 → 2407 lines.

Smoke coverage: wiki pages/graph/stats 200, missing page 404, path
traversal 400, l2 upload 201 + empty-payload 400.
